Check for updates if we are on the update-core
If we are on the update-core screen, check for updates. Otherwise don't. Updates were being triggered at all hours of the day repeatedly - when we really only need to check for updates when the site owner tells it to. This does that by using a transient that expires every 24 hours (so it still can check once per day on its own) - I may change even that though.