Metabox Saving Fix if multiple tabs
If the user was opening several tabs, one after another, the Sessions variable which holds the "old" data (to see if it has changed before saving it), was holding information from different posts - causing it not to save correctly.

This update attaches the session variable to the post id as well - so that we can open many different posts in many different tabs without overwriting the PHP session variable - causing it to save each post correctly.