Field Presets
Field presets are now possible because we can check if a field has ever been saved before. Yes: A Field (not a post).

So a preset value for a field will only show up if the field has never been saved. If the field has been saved, it will use that saved value instead - even if that value is blank.