Only save meta if field value is different
This only saves a meta value if the field value is different, or if it isn't the default value. This change REQUIRES defaults to be provided in code. Meta should be fetched using mp_core_get_post_meta witha  default value set. This doens't apply to checkboxes or repeaters - only to single fields.